services/traffic/requirements.txt
fastapi
uvicorn
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger
prometheus-client
Deprecated